refactor: 프로필이미지 컴포넌트에서 이벤트 핸들러 함수 및 이미지패치/업데이트 함수 분리 - 이벤트핸들러만 모아둔 함수(setAddEventListener, processReduce, removePhoto) - 이미지를 가져와서 업데이트하는 함수(fetchProfileImage, updateProfileImage) Github issue #208
